Break mqtt-broker --> mqtt-edgehub dependency
