+++ mqtt/mqttd/src/broker/bootstrap/edgehub.rs
@@ -14,12 +14,11 @@ use futures_util::{
 use tokio::time;
 use tracing::{info, warn};
 
-use mqtt_broker::{
-    Broker, BrokerBuilder, BrokerSnapshot, MakeEdgeHubPacketProcessor, Server, ServerCertificate,
-};
+use mqtt_broker::{Broker, BrokerBuilder, BrokerSnapshot, Server, ServerCertificate};
 use mqtt_broker_core::{auth::Authorizer, settings::BrokerConfig};
 use mqtt_edgehub::{
     auth::{EdgeHubAuthenticator, EdgeHubAuthorizer, LocalAuthenticator, LocalAuthorizer},
+    connection::MakeEdgeHubPacketProcessor,
     edgelet,
     settings::Settings,
 };
